+++ src/connmgr/connection.rs
@@ -59,9 +59,10 @@ use crate::core::net::{AsyncTcpStream, SocketAddr};
 use crate::core::reactor::Reactor;
 use crate::core::select::{select_2, select_3, select_4, select_option, Select2, Select3, Select4};
 use crate::core::shuffle::random;
+use crate::core::task::{poll_async, CancellationToken};
+use crate::core::time::Timeout;
 use crate::core::waker::RefWakerData;
 use crate::core::zmq::MultipartHeader;
-use crate::future::{poll_async, CancellationToken, Timeout};
 use arrayvec::{ArrayString, ArrayVec};
 use ipnet::IpNet;
 use log::{debug, log, warn, Level};
