+++ src/connmgr/server.rs
@@ -38,12 +38,11 @@ use crate::core::reactor::Reactor;
 use crate::core::select::{
     select_2, select_3, select_6, select_8, select_option, Select2, Select3, Select6, Select8,
 };
+use crate::core::task::{event_wait, yield_to_local_events, CancellationSender, CancellationToken};
+use crate::core::time::Timeout;
 use crate::core::tnetstring;
 use crate::core::waker::RefWakerData;
 use crate::core::zmq::SpecInfo;
-use crate::future::{
-    event_wait, yield_to_local_events, CancellationSender, CancellationToken, Timeout,
-};
 use arrayvec::{ArrayString, ArrayVec};
 use log::{debug, error, info, warn};
 use mio::net::{TcpListener, TcpStream, UnixListener};
