+++ src/connmgr/tls.rs
@@ -18,8 +18,8 @@ use crate::core::event::{self, ReadinessExt};
 use crate::core::io::{AsyncRead, AsyncWrite};
 use crate::core::net::AsyncTcpStream;
 use crate::core::reactor::Registration;
+use crate::core::task::get_reactor;
 use crate::core::waker::{RefWake, RefWaker, RefWakerData};
-use crate::future::get_reactor;
 use arrayvec::ArrayString;
 use log::debug;
 use mio::net::TcpStream;
