+++ src/lib.rs
@@ -25,8 +25,6 @@
 pub mod connmgr;
 pub mod core;
 /// cbindgen:ignore
-pub mod future;
-/// cbindgen:ignore
 pub mod handler;
 /// cbindgen:ignore
 pub mod proxy;
